player and enemy health
added in current and max health for enemies and player.

added scriptable object game events to reduce dependancies and act like a pub/sub.

updated hunter bullet to deal damage when it collides with an object that has health.

Added the UI heart elements to represent the player's health info.