health point hearts on player UI
added heart icons which represent the player's current health and missing heart icons that represent the player's missing health on the screen.

when the player health change event is raised, hearts will either be added or removed depending on if the player lost or gained health.

Changed take damage and restore health functions on player health script to only trigger if the player's health actually changes.